Add hack-data.ini file
This can be used to park any particular data that we don't want to have
to update the core OS for. Right now it is used for a flip-to-hack
blacklist.

Sadly, the GKeyFile parser, unlike the Python .ini parser, does not
allow end-of-line escapes...

In the blacklist we include all browsers, Nautilus, and gnome-software.

https://phabricator.endlessm.com/T24416