Add flip-to-hack whitelist
Intended for Episode 1 only, this whitelist includes all apps that need
to be hackable in Episode 1. In subsequent episodes, all the apps on the
machine should be hackable, so when we fix the issue with the
flip-to-hack button covering up parts of the UI we should empty the
whitelist.

The whitelist is implemented in gnome-shell such that if the key is
empty or not present in the config file, then it will be ignored and all
apps not in the blacklist will be hackable.

https://phabricator.endlessm.com/T24655